Actualized stepthrough
There were some litlle mistakes in the c code comments that tell you the steps of this hack.
First, in step 3. write :

 "cc -g -Wall -o 32bitEFI 32bitEFI.c" 

(and  NOT "cc -g -Wall -o 32bitEFI.c" or you will get a fatal error : No input file)
Also when you run this command you should be IN the folder that contains both the code and the .iso.

Then before step 5 you should actually RUN the program that has just been compiled as "32bitEFI" (not 32bitEFI.c, this is your c script). 

In order to do that in the terminal, type "./32bitEFI". Now the iso has been modified and you shall burn it.